Offline Translator Project
Created the base of a Offline Translator that uses ISO 639-1 language codes in order to translate the user inputted text and then provides the translation in the selected language. The user inputs are written onto a file and the outputs are also written onto a different file.

The project was completed though the use of the translate library from pypi.org developed by Terry.Yin